requirements.txt
colorama==0.4.6
docopt==0.6.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
yarg==0.1.10
requests==2.32.5
python-dotenv==1.2.1
pytest==9.0.2